requirements.txt
azure_storage==0.37.0
beautifulsoup4==4.12.3
fastapi==0.114.2
numpy==2.1.1
pandas==2.2.2
pydantic==2.9.1
python-dotenv==1.0.1
SQLAlchemy==2.0.27
